move windows warning to separate helper
