+++ components/tools/OmeroPy/src/omero/plugins/db.py
@@ -31,7 +31,7 @@ from omero.cli import CLI
 
 from omero_ext.argparse import FileType, SUPPRESS
 
-from omero.plugins.prefs import windows_warning, WINDOWS_WARNING
+from omero.install.windows_warning import windows_warning, WINDOWS_WARNING
 
 from path import path
