+++ components/tools/OmeroPy/src/omero/plugins/node.py
@@ -20,7 +20,7 @@ import signal
 import platform
 from path import path
 
-from omero.plugins.prefs import windows_warning, WINDOWS_WARNING
+from omero.install.windows_warning import windows_warning, WINDOWS_WARNING
 
 HELP = """Control icegridnode.
